+++ src/ConfigManager.cpp
@@ -74,10 +74,8 @@ AppConfig ConfigManager::LoadConfig(const std::wstring& filePath) {
     std::wstring val;
     if (findValue(L"CloseToTray", val)) config.closeToTray = ParseBool(val);
     if (findValue(L"ShowInTaskbar", val)) config.showInTaskbar = ParseBool(val);
-    if (findValue(L"ShowBoundaries", val)) config.showBoundaries = ParseBool(val);
     if (findValue(L"StartWithWindows", val)) config.startWithWindows = ParseBool(val);
     if (findValue(L"WarmTint", val)) config.warmTint = ParseBool(val);
-    if (findValue(L"FocusMode", val)) config.focusMode = ParseBool(val);
     if (findValue(L"IdleDimEnabled", val)) config.idleDimEnabled = ParseBool(val);
     if (findValue(L"IdleMinutes", val)) config.idleMinutes = ParseInt(val);
     if (findValue(L"IdleDimLevel", val)) config.idleDimLevel = ParseInt(val);
@@ -153,10 +151,8 @@ void ConfigManager::SaveConfig(const std::wstring& filePath, const AppConfig& co
     file << L"{\n";
     file << L"  \"CloseToTray\": " << (config.closeToTray ? L"true" : L"false") << L",\n";
     file << L"  \"ShowInTaskbar\": " << (config.showInTaskbar ? L"true" : L"false") << L",\n";
-    file << L"  \"ShowBoundaries\": " << (config.showBoundaries ? L"true" : L"false") << L",\n";
     file << L"  \"StartWithWindows\": " << (config.startWithWindows ? L"true" : L"false") << L",\n";
     file << L"  \"WarmTint\": " << (config.warmTint ? L"true" : L"false") << L",\n";
-    file << L"  \"FocusMode\": " << (config.focusMode ? L"true" : L"false") << L",\n";
     file << L"  \"IdleDimEnabled\": " << (config.idleDimEnabled ? L"true" : L"false") << L",\n";
     file << L"  \"IdleMinutes\": " << config.idleMinutes << L",\n";
     file << L"  \"IdleDimLevel\": " << config.idleDimLevel << L",\n";